backend/requirements.txt
# Core
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1

# PostgreSQL Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# File Processing
pymupdf==1.23.26
python-docx==1.1.0

# AI/NLP
spacy==3.7.4
scikit-learn==1.4.0
numpy==1.26.4

# Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Utilities
aiofiles==23.2.1

